+++ components/comment-rxn.tsx
@@ -1,0 +1,50 @@
+import { Popover } from '@headlessui/react'
+import { FaceSmileIcon } from '@heroicons/react/24/outline'
+
+export const freeReactions = [
+  '➕',
+  '➖',
+  '🤔',
+  '😮',
+  '🥳',
+  '💡',
+  '❓',
+  '🔥',
+  '👏',
+]
+
+export const paidReactions = {
+  '🧡': 1,
+  '🏅': 10,
+  '🏆': 100,
+}
+
+const AddReactionIcon = () => (
+  <div className="relative w-5">
+    <FaceSmileIcon className="h-5 w-5" />
+    <div className="absolute right-0 top-0 rounded-full bg-white px-[0.5px] text-[10px] leading-[10px]">
+      +
+    </div>
+  </div>
+)
+
+export function AddRxn() {
+  return (
+    <Popover className="relative">
+      <Popover.Button className="text-gray-500 hover:text-gray-700 focus:outline-0">
+        <AddReactionIcon />
+      </Popover.Button>
+
+      <Popover.Panel className="absolute z-10">
+        <div className="grid grid-cols-2">
+          <a href="/analytics">Analytics</a>
+          <a href="/engagement">Engagement</a>
+          <a href="/security">Security</a>
+          <a href="/integrations">Integrations</a>
+        </div>
+
+        <img src="/solutions.jpg" alt="" />
+      </Popover.Panel>
+    </Popover>
+  )
+}
